Updated quota to be a floating point value, as this makes elections involving transfer values more accurate, and fixes a bug due to rounding down which was currently present.

+++ Poller/poll_data.py
@@ -30,7 +30,7 @@ class PollResults:
     first_preferences: list[int]
     election_stages: list[tuple[str, list[int], float]]
     votes_per_stage: list[list[float]]
-    quota: int
+    quota: float
 
 
 @dataclass
